requirements.txt
streamlit
openpyxl
python-dotenv
langchain
langchain-community
langchain-openai
faiss-cpu
tenacity
sentence-transformers
openai
transformers
torch